# backend/src/docs/gen/swagger.yaml
basePath: /api
definitions:
  api_example_cognitive.ChatRequest:
    properties:
      context_history:
        type: integer
      max_documents:
        type: integer
      message:
        type: string
      session_id:
        type: integer
      use_rag:
        type: boolean
    required:
    - message
    type: object
...
  github_com_moasq_backend_app_example_cognitive_domain.ChatMessage:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      referenced_docs:
        items:
          type: integer
        type: array
      role:
        $ref: '#/definitions/github_com_moasq_backend_app_example_cognitive_domain.ChatRole'
      session_id:
        type: integer
      tokens_used:
        type: integer
    type: object
  github_com_moasq_backend_app_example_cognitive_domain.ChatResponse:
    properties:
      message:
        $ref: '#/definitions/github_com_moasq_backend_app_example_cognitive_domain.ChatMessage'
      referenced_docs:
        items:
          $ref: '#/definitions/github_com_moasq_backend_app_example_cognitive_domain.SimilarDocument'
        type: array
      session_id:
        type: integer
      tokens_used:
        type: integer
    type: object
  github_com_moasq_backend_app_example_cognitive_domain.ChatRole:
    enum:
    - user
    - assistant
    - system
    type: string
    x-enum-varnames:
    - ChatRoleUser
    - ChatRoleAssistant
    - ChatRoleSystem
  github_com_moasq_backend_app_example_cognitive_domain.SimilarDocument:
    properties:
      chunk_index:
        type: integer
      content_hash:
        type: string
      content_preview:
        type: string
      created_at:
        type: string
      document_id:
        type: integer
      embedding:
        description: 1536 dimensions for OpenAI
        items:
          type: number
        type: array
      id:
        type: integer
      organization_id:
        type: integer
      similarity_score:
        type: number
      updated_at:
        type: string
    type: object
...
  github_com_moasq_backend_pkg_common_errors.HTTPError:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for B2B SaaS Starter.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: B2B SaaS Starter API
  version: "1.0"
paths:
...
  /example_cognitive/chat:
    post:
      consumes:
      - application/json
      description: Sends a message to the AI and gets a response, optionally using
        RAG
      parameters:
      - description: Chat request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_example_cognitive.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_moasq_backend_app_example_cognitive_domain.ChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_common_errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_common_errors.HTTPError'
      summary: Chat with AI
      tags:
      - Cognitive
  /example_cognitive/sessions:
    get:
      description: Lists chat sessions for the current user with pagination
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_common_errors.HTTPError'
      summary: List chat sessions
      tags:
      - Cognitive
  /example_cognitive/sessions/{id}/messages:
    get:
      description: Retrieves all messages for a chat session
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_moasq_backend_app_example_cognitive_domain.ChatMessage'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_common_errors.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_moasq_backend_pkg_common_errors.HTTPError'
      summary: Get session history
      tags:
      - Cognitive
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"